+++ buildpacks/npm/lib/build.sh
@@ -150,13 +150,14 @@ TOML
 clear_cache_on_node_version_change() {
 	local layers_dir=$1
 	local curr_node_version
-	# shellcheck disable=SC2005
-	curr_node_version="$(echo "$(node -v)")"
-	curr_node_version=${curr_node_version:1}
 
-	if [[ "$curr_node_version" != "$PREV_NODE_VERSION" ]]; then
-		info "Deleting cache because node version changed from \"$PREV_NODE_VERSION\" to \"$curr_node_version\""
-		rm -rf "${layers_dir:?}"/*
+	curr_node_version="$(node -v)"
+	curr_node_version=${curr_node_version:1} #to truncate the "v" that is concatedated to version in node -v
+	if [[ ! -z "$PREV_NODE_VERSION" ]]; then
+		if [[ "$curr_node_version" != "$PREV_NODE_VERSION" ]]; then
+			info "Deleting cache because node version changed from \"$PREV_NODE_VERSION\" to \"$curr_node_version\""
+			rm -rf "${layers_dir:?}"/*
+		fi
 	fi
 }
